hukuk-asistan-main/requirements.txt
fastapi
uvicorn[standard]
pydantic-settings
pydantic[email]
loguru
httpx  # n8n'e asenkron HTTP isteği yapmak için
python-dotenv
jinja2
google-generativeai  # Google Gemini AI
# Firestore is included in google-cloud-firestore above
# Google Cloud dependencies
google-cloud-firestore
google-cloud-secret-manager
google-cloud-storage
# Security dependencies
PyJWT
bcrypt
slowapi
cryptography
python-multipart
# Production dependencies
gunicorn
redis
prometheus-client
psutil
# Testing dependencies
pytest
pytest-asyncio
pytest-cov
httpx